% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mae.R
\name{mae}
\alias{mae}
\title{Mean absolute error}
\usage{
mae(actual, predicted, w = NULL, ...)
}
\arguments{
\item{actual}{Observed values.}

\item{predicted}{Predicted values.}

\item{w}{Optional case weights.}

\item{...}{Further arguments passed to \code{weighted_mean}.}
}
\value{
A numeric vector of length one.
}
\description{
Calculates weighted mean absolute error of predicted values.
}
\examples{
mae(1:10, c(1:9, 12))
mae(1:10, c(1:9, 12), w = rep(1, 10))
mae(1:10, c(1:9, 12), w = 1:10)
}
