% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medae.R
\name{medae}
\alias{medae}
\title{Median Absolute Error}
\usage{
medae(actual, predicted, w = NULL, ...)
}
\arguments{
\item{actual}{Observed values.}

\item{predicted}{Predicted values.}

\item{w}{Optional case weights.}

\item{...}{Further arguments passed to \code{weighted_mean}.}
}
\value{
A numeric vector of length one.
}
\description{
Calculates weighted median absolute error of predicted values. The smaller the value, the better.
}
\examples{
medae(1:10, c(2:10, 100))
medae(1:10, c(2:10, 100), w = 1:10)
}
