% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_squared.R
\name{r_squared}
\alias{r_squared}
\title{R-squared}
\usage{
r_squared(actual, predicted, w = NULL, ...)
}
\arguments{
\item{actual}{Observed values.}

\item{predicted}{Predicted values.}

\item{w}{Optional case weights.}

\item{...}{Further arguments passed to \code{weighted_mean}.}
}
\value{
A numeric vector of length one.
}
\description{
Returns weighted R-squared of predicted values.
}
\examples{
r_squared(1:10, c(1, 1:9))
r_squared(1:10, c(1, 1:9), w = rep(1, 10))
r_squared(1:10, c(1, 1:9), w = 1:10)
}
\author{
Michael Mayer, \email{mayermichael79@gmail.com}
}
