% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_squared_gamma.R
\name{r_squared_gamma}
\alias{r_squared_gamma}
\title{Pseudo R-Squared regarding Gamma deviance}
\usage{
r_squared_gamma(actual, predicted, w = NULL, ...)
}
\arguments{
\item{actual}{Observed values.}

\item{predicted}{Predicted values.}

\item{w}{Optional case weights.}

\item{...}{Further arguments passed to \code{r_squared}.}
}
\value{
A numeric vector of length one.
}
\description{
Wrapper to \code{r_squared} with \code{deviance_function = deviance_gamma}.
}
\examples{
r_squared(1:10, c(1:9, 12), w = 1:10, deviance_function = deviance_gamma)
r_squared_gamma(1:10, c(1:9, 12), w = 1:10)
}
\seealso{
\code{\link{r_squared}}.
}
