% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_mean.R
\name{weighted_mean}
\alias{weighted_mean}
\title{Weighted mean that handles NULL weights}
\usage{
weighted_mean(x, w = NULL, ...)
}
\arguments{
\item{x}{Numeric vector.}

\item{w}{Optional case weights.}

\item{...}{Further arguments passed to \code{mean} or \code{weighted.mean}.}
}
\value{
A length-one numeric vector.
}
\description{
Returns weighted mean of numeric vector.
}
\examples{
weighted_mean(1:10)
weighted_mean(1:10, w = NULL)
weighted_mean(1:10, w = 1:10)
}
\author{
Michael Mayer
}
