% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance_poisson.R
\name{deviance_poisson}
\alias{deviance_poisson}
\title{Poisson deviance}
\usage{
deviance_poisson(actual, predicted, w = NULL, ...)
}
\arguments{
\item{actual}{Observed values.}

\item{predicted}{Strictly positive predicted values.}

\item{w}{Optional case weights.}

\item{...}{Further arguments passed to \code{weighted_mean}.}
}
\value{
A numeric vector of length one.
}
\description{
Weighted average of unit Poisson deviance, see e.g. [1]. Special case of Tweedie deviance with Tweedie parameter 1.
}
\examples{
deviance_poisson(0:2, c(0.1, 1, 3))
deviance_poisson(0:2, c(0.1, 1, 3), w = c(1, 1, 1))
deviance_tweedie(0:2, c(0.1, 1, 3), tweedie_p = 1)
deviance_tweedie(0:2, c(0.1, 1, 3), tweedie_p = 1.01)
deviance_poisson(0:2, c(0.1, 1, 3), w = 1:3)
}
\references{
[1] Ohlsson E. and Johansson B. (2015). Non-Life Insurance Pricing with Generalized Linear Models. Springer Nature EN. ISBN 978-3642107900.
}
\seealso{
\code{\link{deviance_tweedie}}.
}
