% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mape.R
\name{mape}
\alias{mape}
\title{Mean absolute percentage error}
\usage{
mape(actual, predicted, w = NULL, ...)
}
\arguments{
\item{actual}{Strictly positive observed values.}

\item{predicted}{Predicted values.}

\item{w}{Optional case weights.}

\item{...}{Further arguments passed to \code{weighted_mean}.}
}
\value{
A numeric vector of length one.
}
\description{
Calculates weighted mean absolute percentage error of predicted values.
}
\examples{
mape(1:10, (1:10)^2)
mape(1:10, (1:10)^2, w = rep(1, 10))
mape(1:10, (1:10)^2, w = 1:10)
}
