% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{mse}
\alias{mse}
\title{Mean-squared error}
\usage{
mse(actual, predicted, w = NULL, ...)
}
\arguments{
\item{actual}{Observed values.}

\item{predicted}{Predicted values.}

\item{w}{Optional case weights.}

\item{...}{Further arguments passed to \code{weighted_mean}.}
}
\value{
A numeric vector of length one.
}
\description{
Calculates weighted mean-squared error of prediction. Equals mean unit normal deviance.
}
\examples{
mse(1:10, c(1:9, 12))
mse(1:10, c(1:9, 12), w = rep(1, 10))
mse(1:10, c(1:9, 12), w = 1:10)
}
\seealso{
\code{\link{rmse}, \link{deviance_normal}}.
}
