% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_median.R
\name{weighted_median}
\alias{weighted_median}
\title{Weighted Median}
\usage{
weighted_median(x, w = NULL, ...)
}
\arguments{
\item{x}{Numeric vector.}

\item{w}{Optional non-negative case weights.}

\item{...}{Further arguments passed to \code{weighted_quantile}.}
}
\description{
Calculates weighted median.
}
\examples{
n <- 20
quantile(seq_len(n), probs = 0.5, type = 4, names = FALSE)
weighted_median(seq_len(n), type = 4)
weighted_median(seq_len(n), w = rep(1, n))
weighted_median(seq_len(n), w = seq_len(n))
}
\seealso{
\code{\link{weighted_quantile}}.
}
