% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_quantile.R
\name{weighted_quantile}
\alias{weighted_quantile}
\title{Weighted Quantiles}
\usage{
weighted_quantile(x, w = NULL, probs = seq(0, 1, 0.25), na.rm = TRUE,
  names = TRUE, ...)
}
\arguments{
\item{x}{Numeric vector.}

\item{w}{Optional non-negative case weights.}

\item{probs}{Vector of probabilities.}

\item{na.rm}{Ignore missing data?}

\item{names}{Return names?}

\item{...}{Further arguments passed to \code{stats::quantile} in the unweighted case. Not used in the weighted case.}
}
\description{
Function copied (and adapted) from spatstat package. Based on linear interpolation of the weighted empirical cdf (like type 4 in stats::quantile).
}
\examples{
n <- 11
quantile(seq_len(n))
weighted_quantile(seq_len(n), type = 4)
weighted_quantile(seq_len(n), w = rep(1, n))
weighted_quantile(seq_len(n), w = seq_len(n))
weighted_quantile(seq_len(n), w = seq_len(n), names = FALSE)
weighted_quantile(seq_len(n), w = seq_len(n), probs = 0.5, names = FALSE)
}
\seealso{
\code{\link{weighted_median}}.
}
